gradio/requirements.txt
gradio
ultralytics
httpx
pillow